build(codegen): fix openapi concurrent jar download
Added a script to  warm up the file-system cache where the
.jar files are downloaded (which contain the OpenAPI generator)

This way the .jar download does not happen concurrently for 26
packages but instead just the one time right before the codegen
script is executed.

IMPORTANT NOTE: If you change the OpenAPI generator version
that is being used by the project from the current v5.2.1 and v6.3.0
then these warm-up scripts must also be updated at the same time
otherwise the same issue that we are fixing here now will creep back
in immediately.

 Fixes #2461

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>