docs(examples): upgrade Angular to v14
1. Upgraded Angular from v13 to v14
2. Also bumped up the Ionic Framework dependencies
as needed because older versions that we were using
were not compatible [1] with Angular v14
3. Also did some of the Cactus -> Cacti renaming
in the front-end code namely the front-page of
both the supply chain and the carbon accounting
app examples.

[1] https://stackoverflow.com/a/72508644

Verbatim copy of the above link's comment:
==================================
The current version of Ionic v6.1.8 is not
compatible with Angular 14. Apparently,
it's been fixed in a dev release but it's not
stable yet. Have a look here
https://github.com/ionic-team/ionic-framework/issues/25353
you can update @ionic/angular to
version 6.1.9-dev.11654275237.1b595be3
and re-run npm install.
Should work fine for now, if you wanna be on the edge.
==================================

Fixes #2377

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>