docs(examples): upgrade Angular to v15
1. Upgraded from Angular v14 to v15 for both GUI packages that use it.
2. Upgraded the Typescript compiler project-wide to 4.8.4
3. Due to 2) There was a change needed in the file
examples/cactus-example-discounted-asset-trade/transaction-fabric.ts
which introduces a constraint on the T type parameter.
4. Disabled the staged files linting because right now ESLint is broken
due to our Typescript version being >=4.8.0 (and upgrading it was not
an option because then it is broken due to ESM issues...)
5. Cleaned out a lot of the unused dev dependencies of the front-end
packages where angular was upgraded. These dependencies were related to
end-to-end browser testing which we haven't set up yet to be executed
because all the front-end packages are just example code that - while
important to test - does not have the highest priority in this sense.
6. Set the rootDir property in the electricity trade example's tsconfig.json
file to the standard value (./) because without it the new tsc compiler
version seems to be putting the compiled source files in a different
location by default (probably an accidental breaking change on the compiler's
part or just a lack of my understanding of how the configuration of it
should work). With the rootDir explicitly specified, it appears to be
compiling the files in the correct location as before.
7. De-duplicated the examples/cactus-example-tcs-huawei/tsconfig.json
file's compilerOptions.tsBuildInfoFile property because it was in collision
with the electricity trade example's property of the same name, leading to
the compiler thinking that it had already compiled the package's code
when it has not.

Fixes #2229

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>