refactor(core-api,verifier-client): use unknown instead of any type
1. Changes the ISocketApiClient and the Verifier type definitions so that
instead of `any` they are heavily relying on `unknown` which helps spotting
bugs in the code where type checks should be done at runtime but currently
aren't.
2. It also helps with spotting breaking changes in the future because if
the test cases stop compiling it means you've done a breaking API change,
but they'll always compile if the parameters and the return values are
using `any` for their own types.

3. I've updated a big chunk of the test cases to have explicit type casts
onto the shapes that are needed for them to compile (without changing
any of the runtime behavior).
4. Also made it so that the json2str function of the DriverCommon
code file now accepts unknown instead of object type which makes one of
the linter warnings go away as well.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>