fix(cmd-api-server): fix CVE-2023-36665 protobufjs try 2
1. Upgraded fabric-network from 2.2.10 to 2.2.18 wherever it was still 2.2.10
2. Upgraded ipfs-http-client project-wide from 51.0.1 to 60.0.1
3. Upgraded @google-cloud/secret-manager from 3.9.0 to 5.0.1

This is the second try at fixing this issue. While it was in review
other commits already snuck back in more older versions of the vulnerable
dependency so before merging this I also add a resolution in the root
package.json file to ensure that the vulnerable version is not used
at all regardless of it being a dependency of a dependency.

[skip ci]

Fixes #2682

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>