docs(examples/cbdc): upgrade web3 from v1.5.2 to v1.10.1
1. This had to be done because of security vulnerabilities in the old version.
2. Originally the robots have attempted to send a pull request with the
same change but it somehow went haywire and upgraded dozens of other
versions in dozens of other packcages not the intended one...
3. So this was manually created to address that bug in GitHub's
dependabot.
4. The original commit message did not mention which vulnerabilities
are  being fixed by it and I also cannot remember the specific ones but
the older versions of web3 were definitely being affected and therefore
it is known to be a good idea what the bot has proposed even though it
couldn't explain itself.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>